% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_id.R
\name{auto_id}
\alias{auto_id}
\title{Automatic classification of sound events in recordings}
\usage{
auto_id(
  model_path,
  updateProgress,
  metadata,
  file_path,
  out_file,
  out_dir,
  save_png = T,
  win_size = 50,
  plot2console = F,
  remove_noise = T,
  recursive = FALSE,
  tx = 1
)
}
\arguments{
\item{model_path}{Path to the fitted model.}

\item{updateProgress}{Progress bar only to be used inside shiny.}

\item{metadata}{Parameters used to create train data for fitting the model.}

\item{file_path}{Path to the folder containing sound recordings.}

\item{out_file}{Character. Name of the file to output the results.
Will be used to name the csv file and the sqlite database.}

\item{out_dir}{Path to the folder where the output results will be stored.}

\item{save_png}{Logical. Should a spectrogram of the classified recordings
with the identified event(s) and respective classification(s) be saved
as png file?}

\item{win_size}{Window size to split recordings in chunks for classification.
One peak per chunk is obtained and classified.}

\item{plot2console}{Logical. Should a spectrogram of the classified
recordings with the identified event(s) and respective classification(s)
be plotted in the console while the analysis is running?
as png file?}

\item{remove_noise}{Logical. TRUE indicates that the model was fitted
with a non-relevant class which will be deleted from the final output.}

\item{recursive}{Logical. FALSE indicates that the recordings are in
a single folder and TRUE indicates that there are recordings
inside subfolders.}

\item{tx}{Time expanded. Only used in recorders specifically intended for
bat recordings. Can take the values "auto" or any numeric value. If the
recording is not time expanded tx must be set to 1 (the default). If it's
time expanded the numeric value corresponding to the time expansion should
be indicated or "auto" should be selected. If tx = "auto" the function
assumes that sampling rates < 50kHz corresponds to
tx = 10 and > 50kHz to tx = 1.}
}
\value{
Nothing.
}
\description{
Applies automatic classification of sound events on a set of
recordings using a fitted model.
}
\details{
Runs a classification task on the recordings of a specified folder
and saves the results of the analysis
}
\author{
Bruno Silva
}
