% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_auto_id_shiny.R
\name{auto_id_shiny}
\alias{auto_id_shiny}
\title{Automatic classification on a set of recordings}
\usage{
auto_id_shiny(
  model_path,
  updateProgress,
  metadata,
  file_path,
  out_file,
  out_dir,
  save_spec = F,
  save_png = T,
  win_size = 50,
  plot2console = F,
  remove_noise = T,
  recursive = FALSE
)
}
\arguments{
\item{model_path}{An object of class 'recording', created with
the 'import_audio' function}

\item{updateProgress}{Progress bar only to be used inside shiny}

\item{metadata}{Parameters used to create train data for fitting the model}

\item{file_path}{Path to the folder containing sound recordings}

\item{out_file}{Character. Name of the file to output the results.
Will be used to name the csv file and the sqlite database}

\item{out_dir}{Path to the folder where the output results will be stored}

\item{save_png}{Logical. Should a spectrogram of the classified recordings
with the identified event(s) and respective classification(s) be saved
as png file?}

\item{win_size}{Window size to split recordings in chunks for classification.
One peak per chunk is obtained and classified}

\item{plot2console}{Logical. Should a spectrogram of the classified
recordings with the identified event(s) and respective classification(s)
be ploted in the console while the analysis is running?
as png file?}

\item{remove_noise}{= T, Logical. TRUE indicates that the model was fitted
with a non-relevant class which will be deleted from the final output}

\item{recursive}{Logical. FALSE indicates that the recordings are in
a single folder and TRUE indicates that there are recordings
inside subfolders}
}
\value{
Nothing
}
\description{
Applies automatic classification on a set of recordings
using a fitted model
}
\details{
Automatic classification on a set of recordings

Runs a classification task on the recordings of a specified folder
and saves the results of the analysis
}
\author{
Bruno Silva
}
