% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduloTrainRetrain_butter_filter.R
\name{butter_filter}
\alias{butter_filter}
\title{Apply a buttterworth filter to a recording}
\usage{
butter_filter(sound_samples, low, high, fs, tx)
}
\arguments{
\item{sound_samples}{Vector. Sound samples}

\item{low}{Minimum frequency in kHz for the butterworth filter}

\item{high}{Maximum frequency in kHz for the butterworth filter}

\item{order}{Butterworth fiter order}
}
\value{
the filtered sound samples
}
\description{
Apply a buttterworth filter to a recording
using the function signal::butter()
}
\details{
Buttterworth filter
}
\author{
Bruno Silva
}
