% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduloCallBD_call_label.R
\name{call_label}
\alias{call_label}
\title{Label calls on a recording}
\usage{
call_label(
  sound_samples,
  file_name,
  fs,
  tx,
  db,
  plot = TRUE,
  win_size = c(13, 5)
)
}
\arguments{
\item{sound_samples}{The sound samples of the recording}

\item{file_name}{Name of the recording}

\item{fs}{Sampling frequency of the recording}

\item{tx}{Expansion factor of the recording}

\item{db}{Full path to the database}

\item{plot}{Logical indicating if the spectrogram is. Should only be set to
FALSE when using the function inside other function that already plots a spectrogram}

\item{win_size}{A two-element vector, c(widht, height), with the dimension in
inches of the spectrogram window}
}
\value{
A dataframe with the file name, the positions of the calls and the
respective labels. The dataframe has as many rows as there are calls
}
\description{
Label calls on a recording by clicking on a spectrogram.
}
\details{
Writes the call positions and labels to a sqlite3 database
}
\author{
Bruno Silva
}
