% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{denoise}
\alias{denoise}
\title{Denoise and standardize a spectrogram matrix}
\usage{
denoise(spec)
}
\arguments{
\item{spec}{Spectrogram matrix}
}
\value{
A denoised spectrogram matrix, with standardized energy values
between 0 and 1
}
\description{
Applies noise removal to a spectrogram matrix, by subtracting
the mean of each frequency column to all rows of that column. Please note that
columns must represent frequency and rows must represent time.Energy values are afterwards standartized
to min = 0 and max = 1 by dividing by maximum energy value of the spectrogram matrix.
}
\author{
Bruno Silva
}
\keyword{internal}
