% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduloTrainRetrain_denoise_spec.R
\name{denoise_spec}
\alias{denoise_spec}
\title{Denoise a spectrogram}
\usage{
denoise_spec(spec, frequency_bin = T, time_bin = F)
}
\arguments{
\item{spec}{Spectrogram matrix}

\item{frequency_bin}{Logical. If TRUE applies denoising by frequency bin}

\item{time_bin}{Logical. If TRUE applies denoising by timebin}
}
\value{
A denoised matrix spectrogram
}
\description{
Applies noise removal to a spectrogram matrix, by subtracting
the mean of the time or/and frequency bin to all pixels of that bin.
}
\author{
Bruno Silva
}
