% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{Calculate mode}
\usage{
get_mode(v)
}
\arguments{
\item{v}{vector}
}
\value{
The mode of the vector
}
\description{
Calculate the mode of a vector
}
\details{
Calculate the mode of a vector
}
\author{
Bruno Silva
}
\keyword{internal}
