% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduloBatID_internal_is.btr.R
\name{is.btr}
\alias{is.btr}
\title{Verify if object is of class "recording"}
\usage{
is.btr(x)
}
\arguments{
\item{x}{Object}
}
\value{
TRUE or FALSE
}
\description{
Verify if object is of class "recording"
}
\details{
Verify if object is of class "recording"
}
\author{
Bruno Silva
}
\keyword{internal}
