% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rc.R
\name{is_rc}
\alias{is_rc}
\title{Check if an object is of class "rc"}
\usage{
is_rc(x)
}
\arguments{
\item{x}{any R object}
}
\value{
TRUE or FALSE
}
\description{
Check if an object is of class "rc"
}
\details{
Check if an object is of class "rc"
}
\author{
Bruno Silva
}
\keyword{internal}
