% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_load2env.R
\name{load2env}
\alias{load2env}
\title{Load a RDATA file to a new environment}
\usage{
load2env(rdata_file, env = new.env())
}
\arguments{
\item{rdata_file}{Character. Path to a .RDATA file}
}
\value{
A new environment with loaded objects
}
\description{
Load a RDATA file to a new environment
}
\details{
Load a RDATA file to a new environment
}
\author{
Bruno Silva
}
\keyword{internal}
