% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduloBatID_peaks.R
\name{peaks}
\alias{peaks}
\title{Detect peaks}
\usage{
peaks(bat_recording, win_size = 40, plot = F)
}
\arguments{
\item{bat_recording}{Object of class "bat_recording"}

\item{win_size}{Window size in ms to search peaks. Sound is divided into
equal size chunks of win_size duration and one peak is identified in each chunk.}

\item{plot}{If TRUE plots the sound_samples with the peaks identified}
}
\value{
Vector with the temporal position of the identified peaks (in samples)
}
\description{
Detect peaks
}
\details{
Detect peaks in sound files
}
\author{
Bruno Silva
}
