% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_peaks2spec.R
\name{peaks2spec}
\alias{peaks2spec}
\title{Sound to spectrogram}
\usage{
peaks2spec(recording, sound_peaks, spec_size = NA, window_length = NA,
frequency_resolution = NA, time_step_size = NA, dynamic_range = NA,
freq_range = NA)
}
\arguments{
\item{recording}{Object of class "rc"}

\item{sound_peaks}{Peaks detected in recording samples}

\item{spec_size}{Spectrogram size in ms}

\item{window_length}{Moving window length to create the spectrogram in ms}

\item{frequency_resolution}{Spectrogram frequency resolution with higher
numbers meaning better resolution. Specifically, for any integer X provided,
1/X the analysis bandwidth (as determined by the number of samples
in the analysis window) will be used. Note that this greatly impacts
processing time, so adjust with care!}

\item{time_step_size}{Moving window step in ms}

\item{dynamic_range}{Threshold of minimum intensity values to show
in the spectrogram}

\item{freq_range}{Frequency range of the spectrogram. Vector with two values,
refering to the minimum and maximum frequency to show in the spectrogram}
}
\value{
an object of class "calls". This object is a list
with the following components:
\itemize{
\item spec_calls -- matrix with one spectrogram per row
\item img_cols -- number of columns in each spectrogram
\item img_rows -- number of rows in each spectrogram
}
}
\description{
Convert sound to spectrogram
}
\details{
Convert sound to spectrogram
}
\author{
Bruno Silva
}
\keyword{internal}
