% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_td.R
\name{plot_td}
\alias{plot_td}
\title{Plot training spectrograms}
\usage{
plot_td(train_data, index)
}
\arguments{
\item{train_data}{Train data object returned by function spectro_calls()}

\item{index}{Vector indicating th index of the spectrograms to plot.}
}
\value{
A plot
}
\description{
Plot training spectrograms
}
\author{
Bruno Silva
}
