% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduloTrainRetrain_internal_r.R
\name{r}
\alias{r}
\title{Rotate a matrix}
\usage{
r(mat)
}
\arguments{
\item{mat}{a matrix}
}
\value{
Rotated matrix
}
\description{
Rotate a matrix
}
\details{
Rotate a matrix
}
\author{
Bruno Silva
}
\keyword{internal}
