% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_r_cw.R
\name{r_cw}
\alias{r_cw}
\title{Rotate a matrix}
\usage{
r_cw(mat)
}
\arguments{
\item{mat}{a matrix}
}
\value{
Rotated matrix
}
\description{
Rotate a matrix 90º clockwise
}
\details{
Rotate a matrix clockwise
}
\author{
Bruno Silva
}
\keyword{internal}
