% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectro_calls.R
\name{spectro_calls}
\alias{spectro_calls}
\title{Generates spectrograms from recording labels.}
\usage{
spectro_calls(files_path, updateProgress, 
db_path, spec_size = NA, window_length = NA, 
frequency_resolution = NA, time_step_size = NA, dynamic_range = NA,
freq_range = NA)
}
\arguments{
\item{files_path}{Path for the folder containing sound recordings}

\item{updateProgress}{Progress bar only to be used inside shiny}

\item{db_path}{Path for the database of recording labels created with the
shinny app provided in the package}

\item{spec_size}{Spectrogram size in ms}

\item{window_length}{Moving window length to create the spectrogram in ms}

\item{frequency_resolution}{Spectrogram frequency resolution. Integer. A value
of 1 corresponds to frequency bins of 1kHz, 2 to frequency bins of 0.5 kHz and
so on}

\item{time_step_size}{Moving window step in ms}

\item{dynamic_range}{Threshold of minimum intensity values to show
in the spectrogram}

\item{freq_range}{Frequency range of the spectrogram. Vector with two values,
refering to the minimum and maximum frequency to show in the spectrogram}
}
\value{
A list with the spectrogram and the respective label
}
\description{
Generates spectrograms from recording's labels for
classification purposes. The spectrogram parameters are user defined
and should be selected depending on the type of sound event to classify. Three
sets of parameters for the spectrograms are provided, with values prepared for
classification of vocalizations of european bat species (one set for all species
and another excluding Rhinolophus species) and birds. Aditional sets of parameters
for other vocalizations or sound events may also be provided by the user.
}
\details{
Generate spectrograms from labels
}
\author{
Bruno Silva
}
