% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduloTrainRetrain_spectro_calls_shiny.R
\name{spectro_calls_shiny}
\alias{spectro_calls_shiny}
\title{Generates spectrograms from recording labels.}
\usage{
spectro_calls(files_path, db_path, parameters)
}
\arguments{
\item{files_path}{Path for the folder containing sound recordings}

\item{db_path}{Path for the database of recording labels created with the
shinny app provided in the package}

\item{parameters}{Parameters for the spectrograms. A list with:
spectrogram size in ms (spec_size - integer),
moving window length used to create the spectrogram in ms (window_length),
spectrogram frequency resolution in kHz (frequency_resolution - integer),
moving window step in ms (time_step_size - positive numeric),
threshold intensity value (dynamic_range - integer),
frequency range of the spectrogram (freq_range - integer vector).
The predefined values are indicated with: "v1", "v2", "v3".}
}
\value{
A list with the spectrogram and the respective label
}
\description{
Generates spectrograms from recording's labels for
classification purposes. The spectrogram parameters are user defined
and should be selected depending on the type of sound event to classify. Three
sets of parameters for the spectrograms are provided, with values prepared for
classification of vocalizations of european bat species (one set for all species
and another excluding Rhinolophus species) and birds. Aditional sets of parameters
for other vocalizations or sound events may also be provided by the user.
}
\details{
Generate spectrograms from labels
}
\author{
Bruno Silva
}
