% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0finished_train_data_process.R
\name{train_data_process}
\alias{train_data_process}
\title{Process train data for Keras format}
\usage{
train_data_process(rdata_list, seed = 1002)
}
\arguments{
\item{rdata_list}{An object created with
the 'spec_calls' function.}

\item{seed}{Integer. Used for the randomization of the observations.}
}
\value{
A list with two components: data_x, an array with the XX data and
data_y, the labels one-hot-encoded.#'
}
\description{
Processes the train data outputed by function 'spec_calls' to
Keras format.
}
\details{
Process train data for Keras format

The spectrogram matrices are converted into an
array (XX train data) and the labels one-hot-encoded (YY train data). The
observations are also randomized.
}
\author{
Bruno Silva
}
