% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_metadata.R
\name{train_metadata}
\alias{train_metadata}
\title{Obtain train metadata to run a fitted model}
\usage{
train_metadata(train_data, save2disk = NA)
}
\arguments{
\item{train_data}{Output of function \link{spectro_calls}.}
}
\value{
A list with the following components:
\itemize{
\item parameters -- parameters of the spectrograms
\item classes -- class names and respective codes
}
}
\description{
Obtain train metadata from
the output of function \link{spectro_calls}. Needed to run a fitted model
}
\author{
Bruno Silva
}
