% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_deploy_model.R
\docType{data}
\name{ui}
\alias{ui}
\title{Start app deploy model}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.
}
\usage{
ui
}
\description{
Starts the app to deploy the model
}
\keyword{datasets}
