% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0finished_internal_add_record.R
\name{add_record}
\alias{add_record}
\title{Add a record to a sqlite database}
\usage{
add_record(path, df)
}
\arguments{
\item{path}{Character. Full path to the database, including name
and extension.}

\item{df}{Dataframe to append to table "labels" in the database.}
}
\value{
Nothing
}
\description{
Append observations with temporal labels (in samples)
of a sound recording to a pre-existing table named "labels" in a sqlite
database created with \code{\link{create_db}}.
}
\details{
Add a record to a sqlite database
}
\author{
Bruno Silva
}
\keyword{internal}
