% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduloCallBD_add_record.R
\name{add_record}
\alias{add_record}
\title{Add a record to a sqlite database}
\usage{
add_record(path, df, table)
}
\arguments{
\item{path}{Full path to the database, including name and extension}

\item{table}{Table name to append dataframe}
}
\value{
Nothing
}
\description{
Append a dataframe with bat call positions of a recording to a
pre-existing table in a sqlite database
}
\details{
Add a record to a sqlite database
}
\author{
Bruno Silva
}
\keyword{internal}
