% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butter_filter.R
\name{butter_filter}
\alias{butter_filter}
\title{Apply a buttterworth filter to sound samples}
\usage{
butter_filter(sound_samples, low = NA, high = NA, fs, tx, order = 12)
}
\arguments{
\item{sound_samples}{Vector. Sound samples to filter}

\item{low}{Minimum frequency in kHz for the butterworth filter}

\item{high}{Maximum frequency in kHz for the butterworth filter}

\item{fs}{Sampling frequency}

\item{tx}{Expanded time factor}

\item{order}{Filter order}
}
\value{
A vector with the filtered sound samples
}
\description{
Apply a buttterworth filter, high pass or/and low pass,
to sound samples. Based on the function \link[signal]{butter}
}
\details{
Buttterworth filter
}
\examples{
sound <- runif(22000, min = -10000, max = 10000) # 1s sound sample
sound_filt <- butter_filter(sound, low = 4, high = 8,
                            fs = 22000, tx = 1, order = 10) # filter sound
Spectrogram(sound, 22000) # original spectrogram
Spectrogram(sound_filt, 22000) # filtered spectrogram
}
\author{
Bruno Silva
}
