% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduloBatID_classify_calls.R
\name{classify_calls}
\alias{classify_calls}
\title{XXXXXXX}
\usage{
classify_calls(calls, sound_peaks, model)
}
\arguments{
\item{calls}{XXXXXXX}

\item{sound_peaks}{XXXXXXX}

\item{model}{XXXXXXX}
}
\value{
XXXXXXXX
}
\description{
XXXXXXX
}
\details{
XXXXXXX
}
\author{
Bruno Silva
}
