% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduloCallBD_create_db.R
\name{create_db}
\alias{create_db}
\title{Create a sqlite3 database}
\usage{
create_db(path, name)
}
\arguments{
\item{path}{Path to the folder where the database will be created}

\item{name}{Name of the database to be created}
}
\value{
Nothing
}
\description{
Create a sqlite3 database with a predefined table in a specific folder,
if a database with the specified name doesn't exist already
}
\details{
A sqlite3 database with the specified name with a predefined table
is created inside the folder
}
\examples{
path <- '~/Projectos/R_packages/development/recLabel/inst/recordings/'
name <- 'test'
create_db(path, name)
}
\author{
Bruno Silva
}
