% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create_db}
\alias{create_db}
\title{Create a sqlite3 database}
\usage{
create_db(path, db_name, table_name = "labels")
}
\arguments{
\item{path}{Path to the folder where the database will be created}

\item{name}{Name of the database to be created}
}
\value{
Nothing
}
\description{
Create a sqlite3 database with a predefined table in a specific folder,
if a database with the specified name doesn't exist already
}
\details{
A sqlite3 database with the specified name with a predefined table
is created inside the folder
}
\examples{
create_db(".//", db_name = "test")
}
\author{
Bruno Silva
}
