% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_noise.R
\name{find_noise}
\alias{find_noise}
\title{Detect energy peaks in non-relevant recordings}
\usage{
find_noise(recording, nmax = 1, plot = F)
}
\arguments{
\item{recording}{Object of class "rc"}

\item{nmax}{Maximum number of peaks to detect in the recording}

\item{plot}{If TRUE plots the sound_samples with the peak(s) identified}
}
\value{
Vector, in samples, with the temporal position of the
identified peak(s)
}
\description{
Detects the temporal position of the number of peaks, events
with most energy, defined in the parameter nmax
}
\details{
Detect energy peaks in non-relevant recordings
}
\author{
Bruno Silva
}
