% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0finished_internal_get_mode.R
\name{get_mode}
\alias{get_mode}
\title{Obtain the mode of a vector}
\usage{
get_mode(v)
}
\arguments{
\item{v}{vector}
}
\value{
The mode of the vector
}
\description{
Obtain the mode of a vector
}
\details{
Obtain the mode of a vector
}
\author{
Bruno Silva
}
\keyword{internal}
