% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_is_even.R
\name{is_even}
\alias{is_even}
\title{Check if number is even}
\usage{
is_even(x)
}
\arguments{
\item{x}{a number}
}
\value{
TRUE if number is even, FALSE if number is odd
}
\description{
Check if number is even
}
\details{
Check if number is even
}
\author{
Bruno Silva
}
\keyword{internal}
