% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2samples.R
\name{ms2samples}
\alias{ms2samples}
\title{Convert between time and samples in sound files}
\usage{
ms2samples(value, fs = 300000, tx = 1, inv = F)
}
\arguments{
\item{value}{time in ms or number of samples}

\item{fs}{sampling frequency in samples per second}

\item{tx}{time expansion factor}

\item{inv}{if TRUE converts time to samples.
If FALSE converts samples to time}
}
\value{
if TRUE returns number of samples, if FALSE
returns time in ms
}
\description{
Convert time to samples or vice-versa in sound files
}
\details{
Convert between time and samples in sound files
}
\author{
Bruno Silva
}
