% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0finished_ms2samples.R
\name{ms2samples}
\alias{ms2samples}
\title{Convert between time and number of samples in sound files}
\usage{
ms2samples(value, fs = 300000, tx = 1, inv = FALSE)
}
\arguments{
\item{value}{Numeric representing time in ms or number of samples.}

\item{fs}{Integer. The sampling frequency in samples per second.}

\item{tx}{Integer. Time expansion factor.}

\item{inv}{Logical. If TRUE converts time to number of samples, if FALSE
number of samples to time.}
}
\value{
if TRUE returns number of samples, if FALSE
returns time in ms
}
\description{
Convert time to number of samples or vice-versa
in sound files.
}
\details{
Convert between time and number of samples in sound files
}
\examples{
ms2samples(150000, fs = 300000, tx = 1, inv = FALSE)
ms2samples(100, fs = 300000, tx = 1, inv = TRUE)
}
\author{
Bruno Silva
}
