% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0finished_peaks.R
\name{peaks}
\alias{peaks}
\title{Detect energy peaks in recordings}
\usage{
peaks(recording, win_size = 40, plot = F)
}
\arguments{
\item{recording}{Object of class "rc"}

\item{win_size}{Window size in ms to search peaks. Sound is divided into
equal size chunks of win_size duration and one peak is identified
in each chunk.}

\item{plot}{If TRUE plots the sound_samples with the peaks identified}
}
\value{
Vector with the temporal position of the identified peaks in samples
}
\description{
Splits the recording in segments using a pre-defined window size
and detects the temporal position of the peak (event with most energy) in
each window.
}
\details{
Detect peaks in recordings
}
\author{
Bruno Silva
}
\keyword{internal}
