% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduloTrainRetrain_peaks2spec.R
\name{peaks2spec}
\alias{peaks2spec}
\title{Sound to spectrogram}
\usage{
peaks2spec(bat_recording, sound_peaks, frequency_bin = T,
time_bin = F,version = "v1")
}
\arguments{
\item{bat_recording}{Object of class "bat_recording"}

\item{sound_peaks}{Peaks detected in recording samples}

\item{frequency_bin}{If TRUE filter by frequency mean subtraction}

\item{time_bin}{If TRUE filter by time mean subtraction}

\item{version}{Set of parameters for the analysis.}
}
\value{
an object of class "calls". This object is a list
with the following components:
\itemize{
\item spec_calls -- matrix with one spectrogram per row
\item img_cols -- number of columns in each spectrogram
\item img_rows -- number of rows in each spectrogram
}
}
\description{
Convert sound to spectrogram
}
\details{
Convert sound to spectrogram
}
\author{
Bruno Silva
}
