% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_output_shiny.R
\name{save_output_shiny}
\alias{save_output_shiny}
\title{Save output}
\usage{
save_output_shiny(output, recording, out_file = NA, png_file = NA,
plot2console = F, recursive = FALSE)
}
\arguments{
\item{output}{Object of class "tidy_output"}

\item{recording}{Object of class "recording"}

\item{out_file}{Character. Name of databse file to write analysis results. If
NA, no analysis results file is created}

\item{png_file}{Character. Name of file for plotting spectrogram with call
labels. If NA, no spectrogram file is created}

\item{plot2console}{If TRUE plots the spectrogram with the peaks identified
to the console. Please be advised that if TRUE the analysis takes
considerably more time.}

\item{recursive}{Logical. Is the analysis beeing processed in recursive mode?}
}
\value{
Nothing
}
\description{
Writes to csv and sqlite database the results of the
classification.
}
\details{
Output analysis results
}
\author{
Bruno Silva
}
\keyword{internal}
