% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduloCallBD_spectrogram.R
\name{Spectrogram}
\alias{Spectrogram}
\title{Spectrogram}
\usage{
Spectrogram(
  Audio,
  SamplingFrequency = NULL,
  WindowLength = 5,
  FrequencyResolution = 4,
  TimeStepSize = NULL,
  nTimeSteps = NULL,
  Preemphasis = TRUE,
  DynamicRange = 70,
  Omit0Frequency = FALSE,
  WindowType = "kaiser",
  WindowParameter = NULL,
  plot = TRUE,
  PlotFast = TRUE,
  add = FALSE,
  col = NULL,
  xlim = NULL,
  ylim = NULL,
  main = "",
  xlab = "Time (ms)",
  ylab = "Frequency (kHz)",
  norm = 140
)
}
\arguments{
\item{Audio}{An object of class 'recording', created with the 'import_audio' function}
}
\value{
If plot=FALSE, returns a spectrogram matrix, with the parameters used in the analysis as attributes.
If plot=TRUE a plot of the spectrogram is produced but the matrix is not exported.
}
\description{
Plots a spectrogram or returns a spectrogram matrix
}
\author{
Bruno Silva
}
