% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_spectrogram.R
\name{Spectrogram}
\alias{Spectrogram}
\title{Spectrogram}
\usage{
Spectrogram(Audio,
             SamplingFrequency=NULL,
             WindowLength = 5,
             FrequencyResolution = 4,
             TimeStepSize = NULL,
             nTimeSteps = NULL,
             Preemphasis = TRUE,
             DynamicRange = 70,
             Omit0Frequency = FALSE,
             WindowType = "kaiser",
             WindowParameter = NULL,
             plot = TRUE,
             PlotFast = TRUE,
             add = FALSE,
             col = NULL,
             xlim = NULL,
             ylim = NULL,
             main = "",
             xlab = "Time (ms)",
             ylab = "Frequency (kHz)",
             norm = 140)
}
\arguments{
\item{Audio}{Object of class "sound", "Wave", "Sample", "audioSample" or
a numeric vector (representing a sequence of samples taken from a sound wave)}

\item{SamplingFrequency}{The sampling frequency/rate of the sound in Hertz.
Only needed if Audio parameter is a numeric vector}

\item{WindowLength}{The desired length in ms of the analysis window used
to create the spectrogram}

\item{FrequencyResolution}{Integer, with higher numbers meaning
better resolution. Specifically, for any integer X provided, 1/X the
analysis bandwidth (as determined by the number of samples in the
analysis window) will be used. Note that this greatly impacts the
processing time, so adjust with care!}

\item{TimeStepSize}{Number of milliseconds that the window will be moved
for each adjacent analysis}

\item{nTimeSteps}{The overall total number of time steps. Only needed if
TimeStepSize is left NULL}

\item{Preemphasis}{Logical. Should pre-emphasis be applied
(to all frequency regions, i.e. with a dummy frequency cutoff of 0)? In other
words, should the spectral slope at all frequencies increase by 6 dB
per octave using a single-pole filter?}

\item{DynamicRange}{Values less than this many dB below the maximum are
'clipped' to that value. If this is set to NULL, no clipping occurs.}

\item{Omit0Frequency}{Logical. As the frequency band at 0 Hz is usually at
very low values, choose TRUE to omit this frequency band}

\item{WindowType}{A character string indicating the desired type of window
function to be applied to the signal. Can be one of: "rectangular",
"blackman", "hanning", "hamming", "cosine", "bartlett", "gaussian", "kaiser"}

\item{WindowParameter}{Numeric. Only relevant if the WindowType is set to
"gaussian" or "kaiser"; it will be ignored in all other cases.}

\item{plot}{Logical. Whether the spectrogram should be plotted or not. If
FALSE, no spectrogram is plotted, and instead, a matrix is returned
containing the magnitude at each bin center (time across the rows and
frequency across the columns)}

\item{PlotFast}{Logical. If set to FALSE, the filled.contour() function will
be used. This produces much better looking graphics (which is best
for publications), but takes considerably longer to plot. If set to TRUE,
the image() function will be used instead, with 'useRaster' set to TRUE.}

\item{add}{Logical. Determines whether an entirely new plot is
drawn (with all the annotation) or whether just the core image is drawn}

\item{col}{Color map to use for the plot}

\item{xlim}{X axis (representing time) limit}

\item{ylim}{Y axis (representing frequency) limit}

\item{main}{Main title}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{norm}{Maximum intensity for spectrogram normalization (in dB)}
}
\value{
If plot=FALSE, returns a spectrogram matrix, with the parameters
used in the analysis as attributes. If plot=TRUE a plot of the spectrogram
is produced but the matrix is not exported.
}
\description{
Plots a spectrogram or returns a spectrogram matrix
}
\details{
Computes a spectrogram which can be ploted or exported as a matrix
}
\keyword{internal}
