% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduloBatID_tidy_output.R
\name{tidy_output}
\alias{tidy_output}
\title{Format the classification table}
\usage{
tidy_output(output, class_labels, min_dist = 40, remove_noise = T, multiple_species = F)
}
\arguments{
\item{output}{output of function classify_calls()}

\item{class_labels}{vector of character labels for the categories.
Must be in the same order as the numeric equivalent. For instance,
if the numeric categories are 0 = noise and 1 = bat, the vector must
be c("noise", "bat")}

\item{min_dist}{minimum distance in ms between consecutive calls.}

\item{remove_noise}{if TRUE remove the peaks identified as noise
from the formated table}

\item{multiple_species}{if FALSE only the calls belonging to the
most represented species are kept in the formated table. If TRUE all
calls classified as bats are returned in the formated table}
}
\value{
A formated classification table
}
\description{
Format the classification table
for easier interpretation
}
\details{
This function eliminates calls: 1) classified as noise,
and 2) closer than a specific threshold. It
then converts the numeric classes to text labels
}
\author{
Bruno Silva
}
